Move shgeom function to shgeom.py
