+++ qpsolvers_benchmark/utils.py
@@ -23,8 +23,6 @@ import platform
 from importlib import metadata
 from typing import Set
 
-import numpy as np
-
 from .spdlog import logging
 
 try:
@@ -34,33 +32,6 @@ except ImportError:
     logging.warning(
         "Run ``pip install py-cpuinfo`` for more accurate CPU info"
     )
-
-
-def shgeom(v: np.ndarray, sh: float) -> float:
-    """
-    `Shifted geometric mean <http://plato.asu.edu/ftp/shgeom.html>`_.
-
-    Args:
-        v: Nonnegative values.
-        sh: Shift parameter.
-
-    Note:
-        The mean is computed as exponential of sum of logs to avoid memory
-        overflows. This is common practice.
-
-    Notes:
-        Quoting from `A Note on #fairbenchmarking
-        <https://community.fico.com/s/blog-post/a5Q2E000000Dt0JUAS/fico1421>`_:
-        "The geometric mean of n numbers is defined as the n-th root of their
-        product. For the shifted geometric mean, a positive shift value s is
-        added to each of the numbers before multiplying them and subtracted
-        from the root afterwards. Shifted geometric means have the advantage to
-        neither be compromised by very large outliers (in contrast to
-        arithmetic means) nor by very small outliers (in contrast to geometric
-        means)."
-    """
-    assert (v >= 0.0).all() and sh >= 1.0
-    return np.exp(np.sum(np.log(v + sh)) / len(v)) - sh
 
 
 def get_cpu_info():
